// src/lib/api/generated/models/GoldSourcePlayerProfile.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * PolystirolHub Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Player profile
 * @export
 * @interface GoldSourcePlayerProfile
 */
export interface GoldSourcePlayerProfile {
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    steamId: any | null;
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    name: any | null;
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    registered: any | null;
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    totalPlaytime: any | null;
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    totalKills: any | null;
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    totalDeaths: any | null;
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    totalHeadshots: any | null;
    /**
     * 
     * @type {any}
     * @memberof GoldSourcePlayerProfile
     */
    serversPlayed: any | null;
}

/**
 * Check if a given object implements the GoldSourcePlayerProfile interface.
 */
export function instanceOfGoldSourcePlayerProfile(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "steamId" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "registered" in value;
    isInstance = isInstance && "totalPlaytime" in value;
    isInstance = isInstance && "totalKills" in value;
    isInstance = isInstance && "totalDeaths" in value;
    isInstance = isInstance && "totalHeadshots" in value;
    isInstance = isInstance && "serversPlayed" in value;

    return isInstance;
}

export function GoldSourcePlayerProfileFromJSON(json: any): GoldSourcePlayerProfile {
    return GoldSourcePlayerProfileFromJSONTyped(json, false);
}

export function GoldSourcePlayerProfileFromJSONTyped(json: any, ignoreDiscriminator: boolean): GoldSourcePlayerProfile {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'steamId': json['steam_id'],
        'name': json['name'],
        'registered': json['registered'],
        'totalPlaytime': json['total_playtime'],
        'totalKills': json['total_kills'],
        'totalDeaths': json['total_deaths'],
        'totalHeadshots': json['total_headshots'],
        'serversPlayed': json['servers_played'],
    };
}

export function GoldSourcePlayerProfileToJSON(value?: GoldSourcePlayerProfile | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'steam_id': value.steamId,
        'name': value.name,
        'registered': value.registered,
        'total_playtime': value.totalPlaytime,
        'total_kills': value.totalKills,
        'total_deaths': value.totalDeaths,
        'total_headshots': value.totalHeadshots,
        'servers_played': value.serversPlayed,
    };
}